+++ model/mmwave-ue-mac.h
@@ -1,43 +1,35 @@
 /* -*-  Mode: C++; c-file-style: "gnu"; indent-tabs-mode:nil; -*- */
 /*
- *   Copyright (c) 2011 Centre Tecnologic de Telecomunicacions de Catalunya (CTTC)
- *   Copyright (c) 2015, NYU WIRELESS, Tandon School of Engineering, New York University
- *
- *   This program is free software; you can redistribute it and/or modify
- *   it under the terms of the GNU General Public License version 2 as
- *   published by the Free Software Foundation;
- *
- *   This program is distributed in the hope that it will be useful,
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *   GNU General Public License for more details.
- *
- *   You should have received a copy of the GNU General Public License
- *   along with this program; if not, write to the Free Software
- *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- *
- *   Author: Marco Miozzo <marco.miozzo@cttc.es>
- *           Nicola Baldo  <nbaldo@cttc.es>
- *
- *   Modified by: Marco Mezzavilla < mezzavilla@nyu.edu>
- *                Sourjya Dutta <sdutta@nyu.edu>
- *                Russell Ford <russell.ford@nyu.edu>
- *                Menglei Zhang <menglei@nyu.edu>
- */
+*   Copyright (c) 2011 Centre Tecnologic de Telecomunicacions de Catalunya (CTTC)
+*   Copyright (c) 2015 NYU WIRELESS, Tandon School of Engineering, New York University
+*   Copyright (c) 2019 Centre Tecnologic de Telecomunicacions de Catalunya (CTTC)
+*
+*   This program is free software; you can redistribute it and/or modify
+*   it under the terms of the GNU General Public License version 2 as
+*   published by the Free Software Foundation;
+*
+*   This program is distributed in the hope that it will be useful,
+*   but WITHOUT ANY WARRANTY; without even the implied warranty of
+*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+*   GNU General Public License for more details.
+*
+*   You should have received a copy of the GNU General Public License
+*   along with this program; if not, write to the Free Software
+*   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+*
+*/
 
+#ifndef MMWAVE_UE_MAC_H
+#define MMWAVE_UE_MAC_H
 
-
-#ifndef SRC_MMWAVE_MODEL_MMWAVE_UE_MAC_H_
-#define SRC_MMWAVE_MODEL_MMWAVE_UE_MAC_H_
 
 #include "mmwave-mac.h"
+#include "mmwave-phy-mac-common.h"
+#include "mmwave-phy-sap.h"
 #include <ns3/lte-ue-cmac-sap.h>
-#include <ns3/lte-mac-sap.h>
-#include <ns3/lte-radio-bearer-tag.h>
-#include <queue>
+#include <ns3/lte-ccm-mac-sap.h>
 
 namespace ns3 {
-class MmWaveControlMessage;
 
 class MmWaveUeMac : public Object
 {
@@ -46,34 +38,102 @@ class MmWaveUeMac : public Object
   friend class MacUeMemberPhySapUser;
 
 public:
+  /**
+   * \brief Get the Type id
+   * @return the type id
+   */
   static TypeId GetTypeId (void);
 
+  /**
+   * \brief MmWaveUeMac constructor
+   */
   MmWaveUeMac (void);
-  ~MmWaveUeMac (void);
-  virtual void DoDispose (void);
+  /**
+    * \brief Deconstructor
+    */
+  ~MmWaveUeMac (void) override;
 
+  /**
+   * \brief Set the C MAC SAP user (AKA the RRC representation for the MAC)
+   * \param s the SAP pointer
+   */
   void  SetUeCmacSapUser (LteUeCmacSapUser* s);
+
+  /**
+   * \brief Get the C MAC SAP provider (AKA the MAC representation for the RRC)
+   * \return  C MAC SAP provider (AKA the MAC representation for the RRC)
+   */
   LteUeCmacSapProvider*  GetUeCmacSapProvider (void);
+
+  /**
+   * \brief Get the Mac SAP provider (AKA the MAC representation for the RLC)
+   * \return the Mac SAP provider (AKA the MAC representation for the RLC)
+   */
   LteMacSapProvider*  GetUeMacSapProvider (void);
 
-  void SetConfigurationParameters (Ptr<MmWavePhyMacCommon> ptrConfig);
-  Ptr<MmWavePhyMacCommon> GetConfigurationParameters (void) const;
-
-  void DoSlotIndication (SfnSf sfn);
-
+  /**
+   * \brief Get the PHY SAP User (AKA the MAC representation for the PHY)
+   * \return the PHY SAP User (AKA the MAC representation for the PHY)
+   */
   MmWaveUePhySapUser* GetPhySapUser ();
+
+  /**
+   * \brief Set PHY SAP provider (AKA the PHY representation for the MAC)
+   * \param ptr the PHY SAP provider (AKA the PHY representation for the MAC)
+   */
   void SetPhySapProvider (MmWavePhySapProvider* ptr);
 
-  void RecvRaResponse (BuildRarListElement_s raResponse);
+  /**
+   * \brief Set the configuration parameter for the mac
+   *
+   * Done only once in the helper (no handover)
+   *
+   * \param ptrConfig Config
+   */
+  void SetConfigurationParameters (const Ptr<MmWavePhyMacCommon> &ptrConfig);
 
-  virtual void SetRnti (uint16_t);
+  /**
+   *  TracedCallback signature for Ue Mac Received Control Messages.
+   *
+   * \param [in] frame Frame number.
+   * \param [in] subframe Subframe number.
+   * \param [in] slot number.
+   * \param [in] VarTti
+   * \param [in] ccId
+   * \param [in] pointer to msg to get the msg type
+   */
+  typedef void (* RxedUeMacCtrlMsgsTracedCallback)
+    (const SfnSf sfnSf, const uint16_t rnti, const uint8_t ccId, Ptr<MmWaveControlMessage>);
+
+  /**
+   *  TracedCallback signature for Ue Mac Transmitted Control Messages.
+   *
+   * \param [in] frame Frame number.
+   * \param [in] subframe Subframe number.
+   * \param [in] slot number.
+   * \param [in] VarTti
+   * \param [in] ccId
+   * \param [in] pointer to msg to get the msg type
+   */
+  typedef void (* TxedUeMacCtrlMsgsTracedCallback)
+    (const SfnSf sfnSf, const uint16_t rnti, const uint8_t ccId, Ptr<MmWaveControlMessage>);
+
+
+
+protected:
+  // From object
+  virtual void DoDispose (void) override;
 
 private:
+  void RecvRaResponse (BuildRarListElement_s raResponse);
+  void SetRnti (uint16_t);
+  void DoSlotIndication (SfnSf sfn);
+
   /**
    * \brief Get the total size of the RLC buffers.
    * \return The number of bytes that are in the RLC buffers
    */
-  uint32_t GetTotalBufSize () const;
+  uint32_t GetTotalBufSize () const __attribute__((warn_unused_result));
 
   void SendSR () const;
   void DoTransmitPdu (LteMacSapProvider::TransmitPduParameters params);
@@ -110,23 +170,20 @@ private:
 
   std::map<uint32_t, struct MacPduInfo>::iterator AddToMacPduMap (const std::shared_ptr<DciInfoElementTdma> & dci, unsigned activeLcs);
 
+private:
   Ptr<MmWavePhyMacCommon> m_phyMacConfig;
 
   LteUeCmacSapUser* m_cmacSapUser;
   LteUeCmacSapProvider* m_cmacSapProvider;
-
-  SlotAllocInfo m_DataTxAllocationList;
-
   MmWavePhySapProvider* m_phySapProvider;
   MmWaveUePhySapUser* m_phySapUser;
   LteMacSapProvider* m_macSapProvider;
 
-  uint16_t m_frameNum;
-  uint8_t m_subframeNum;
-  uint16_t m_slotNum;
-  uint8_t m_varTtiNum;
+  uint16_t m_frameNum {0};
+  uint8_t m_subframeNum {0};
+  uint16_t m_slotNum {0};
+  uint8_t m_varTtiNum {0};
 
-  //uint8_t m_tbUid;
   std::map<uint32_t, struct MacPduInfo> m_macPduMap;
 
   std::map <uint8_t, LteMacSapProvider::ReportBufferStatusParameters> m_ulBsrReceived;   // BSR received from RLC (the last one)
@@ -134,31 +191,16 @@ private:
   /**
    * \brief States for the SR/BSR mechanism.
    *
-   * The SR/BSR mechanism is composed by two components: a variable in which
-   * it is saved the state (INACTIVE/ACTIVE) and m_bsrReservedSpace, a queue
-   * of Sfnsf (frame - subframe - slot - symbol start) in which is possible to
-   * send a BSR message.
+   * The SR/BSR mechanism is based on a variable in which
+   * it is saved the state (INACTIVE/ACTIVE).
    *
    * The machine is starting from the INACTIVE state. When the RLC notifies
    * to MAC that there are new bytes in its queue (DoReportBufferStatus()),
    * if the machine is in INACTIVE state, it enters the ACTIVE state.
-   * Entering the ACTIVE state means to send a SR, which will be sent in UL CTRL
-   * symbol of the same slot the data is received. Being already in the ACTIVE
-   * state means that the BSR will be sent only in the reserved Sfnsf,
-   * that are maintained as queue in the variable m_bsrReservedSpace.
-   *
-   * When the UE receive a message of type UL_BSR_GRANT to himself, then
-   * the current slot is saved in the m_bsrReservedSpace
-   * queue. It means that the UE, in the current slot, will send a BSR
-   * in the UL CTRL symbol.
-   *
-   * When the UE receive a DCI UL grant, it saves the Sfnsf in which it will
-   * send the data (depending on the K2) as allowed in the m_bsrReservedSpace
-   * queue. In this way, along with the data, a BSR will be sent.
-   *
-   * At the start of each VarTti (each different DCI) the PHY is calling
-   * DoSlotIndication. In this function, if the Sfnsf at the front of the Sfnsf
-   * queue is true, then the BSR is sent only if we have data in the queue.
+   * Entering the ACTIVE state means to send a SR, which is enqueued in the PHY layer.
+   * It will suffer slots of CTRL latency. If the state is already ACTIVE, then
+   * the BSR can be sent in the same slot as data. It means that the MAC prepares
+   * together the data and the BSR.
    *
    * If the BSR is not sent (we don't have any data in the queue) and we don't
    * have any more reserved space to send BSR, then the state goes back to the
@@ -167,15 +209,14 @@ private:
   enum SrBsrMachine : uint8_t
   {
     INACTIVE,    //!< no SR nor BSR.. initial state
+    TO_SEND,     //!< We have to send the BSR when possible
     ACTIVE       //!< SR or BSR sent; now the source of information is the vector m_bsrReservedSpace
   };
-  std::queue<SfnSf> m_bsrReservedSpace;    //!< The SfnSf(s) in which the UE has reserved space for BSR
   SrBsrMachine m_srState {INACTIVE};       //!< Current state for the SR/BSR machine.
-  bool m_stdBsr {false};                   //!< 3GPP std BSR/SR or old method.
 
   Ptr<UniformRandomVariable> m_raPreambleUniformVariable;
-  uint8_t m_raPreambleId;
-  uint8_t m_raRnti;
+  uint8_t m_raPreambleId {0};
+  uint8_t m_raRnti {0};
 
   struct UlHarqProcessInfo
   {
@@ -196,14 +237,27 @@ private:
   };
 
   std::map <uint8_t, LcInfo> m_lcInfoMap;
-  uint16_t m_rnti;
+  uint16_t m_rnti {0};
 
-  bool m_waitingForRaResponse;
+  bool m_waitingForRaResponse {true};
   static uint8_t g_raPreambleId;
 
+  /**
+   * Trace information regarding Ue MAC Received Control Messages
+   * Frame number, Subframe number, slot, VarTtti, rnti, ccId,
+   * pointer to message in order to get the msg type
+   */
+  TracedCallback<SfnSf, uint16_t, uint8_t, Ptr<MmWaveControlMessage>> m_macRxedCtrlMsgsTrace;
+
+  /**
+   * Trace information regarding Ue MAC Transmitted Control Messages
+   * Frame number, Subframe number, slot, VarTtti, rnti, ccId,
+   * pointer to message in order to get the msg type
+   */
+  TracedCallback<SfnSf, uint16_t, uint8_t, Ptr<MmWaveControlMessage>> m_macTxedCtrlMsgsTrace;
 };
 
 }
 
 
-#endif /* SRC_MMWAVE_MODEL_MMWAVE_UE_MAC_H_ */
+#endif /* MMWAVE_UE_MAC_H */